Created model handler class
