+++ uiwiz/ui.py
@@ -31,3 +31,4 @@ from uiwiz.elements.toast import Toast as toast
 from uiwiz.elements.toggle import Toggle as toggle
 from uiwiz.elements.upload import Upload as upload
 from uiwiz.show import show
+from uiwiz.model_handler import ModelForm as modelForm
